# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Felhasználó regisztrálása",
                "description": "Új felhasználó regisztrálása a rendszerbe.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Regisztrációs adatok",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "nev": "Kovács Balázs",
                                    "email": "kovibali@gmail.com",
                                    "jelszo": "kovibali",
                                    "jelszo_megerositese": "securepassword",
                                    "role": "admin",
                                    "telefonszam": "123456789",
                                    "szemelyi_szam": "12345678",
                                    "szuletesi_datum": "1990-01-01",
                                    "ceg": true,
                                    "cegnev": "Acme Inc.",
                                    "ceg_tipus": "Kft.",
                                    "ado_szam": "123456789",
                                    "bankszamlaszam": "12345678901234567890123456",
                                    "orszag": "Magyarország",
                                    "iranyitoszam": "1234",
                                    "varos": "Budapest",
                                    "utca": "Kossuth utca",
                                    "utca_jellege": "utca",
                                    "hazszam": "10",
                                    "epulet": "A",
                                    "lepcsohaz": "1",
                                    "emelet": "2",
                                    "ajto": "3"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Érvénytelen felhasználó adatok"
                    },
                    "422": {
                        "description": "Érvénytelen regisztrációs adatok"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Bejelentkezés a rendszerbe",
                "operationId": "login",
                "requestBody": {
                    "description": "Bejelentkezési adatok",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "examples": {
                                "Kovacs": {
                                    "summary": "Kovács Balázs",
                                    "value": {
                                        "email": "kovibali@gmail.com",
                                        "jelszo": "kovibali"
                                    }
                                },
                                "Teszt Elek": {
                                    "summary": "Kovács Balázs",
                                    "value": {
                                        "email": "tesztelek@tesztelek.hu",
                                        "jelszo": "tesztelek"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Érvénytelen felhasználói név vagy jelszó"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}